Convert pathlib Path into str before passing to DataLad interfaces
It used to work but now (with 0.14.4) it does not. I did not track which
version changed behavior, but assuming that we never (in datalad) promised
to support and auto cast paths etc into str, not sure if will be fixed
promptly.

Unfortunately with this changes script still does not work for me -- but
it smells to be something about authentication, as if NDA no longer
provides access to those old S3 URLs "I used to know"